requirements.txt
fastapi[standard]
firebase-admin
PyJWT
passlib[bcrypt]
python-dotenv
pytest
python-socketio[asyncio]
httpx
onesignal-python-api
resend
pandas
numpy
scikit-learn
pydantic-ai-slim[openai]
fpdf2 
matplotlib >=3.8.0
